added login component
We need to embed login component to all other components in an ngIf, so that if user is not signed in, then login component shows, else the other component shows. I am trying to think of a way that we only have to embed login component to one file instead of all of them. Any ideas?